app/backend/requirements.txt
flask
flask-cors
python-dotenv
plaid-python
mastercard-oauth1-signer
mastercard-client-encryption
